requirements.txt
numpy==1.23.1
pillow